move method out of commander.ts
