+++ governance/multisig_wh_message_builder/package.json
@@ -22,7 +22,7 @@
     "prepublishOnly": "npm run build && npm test && npm run lint",
     "preversion": "npm run lint",
     "version": "npm run format && git add -A src",
-    "start": "node lib/index.js"
+    "start": "node lib/commander.js"
   },
   "keywords": [
     "pyth",
